@@ -17,15 +17,15 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "illuminate/auth": "^8.0",
-        "illuminate/events": "^8.0",
-        "illuminate/session": "^8.0",
-        "illuminate/support": "^8.0"
+        "php": "^8.0",
+        "illuminate/auth": "^8.0|^9.0",
+        "illuminate/events": "^8.0|^9.0",
+        "illuminate/session": "^8.0|^9.0",
+        "illuminate/support": "^8.0|^9.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.16",
-        "orchestra/testbench": "^6.0",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "orchestra/testbench": "^6.0|^7.0",
         "phpunit/phpunit": "^9.3"
     },
     "autoload": {
