@@ -6,24 +6,16 @@ interface AuthTimeout
 {
     /**
      * Initialize the timeout session when no has been set yet.
-     *
-     * @return void;
      */
-    public function init();
+    public function init(): void;
 
     /**
      * Check if a user has timed out. Returns `true` if not yet timed out and `false` otherwise.
-     *
-     * @param mixed $guard
-     *
-     * @return bool
      */
-    public function check($guard = null);
+    public function check(string $guard = null): bool;
 
     /**
      * Reset the user's timeout session.
-     *
-     * @return void
      */
-    public function reset();
+    public function reset(): void;
 }
